refactor(ui): improve type safety and component rendering
- Add explicit type casting for UI component dimensions
- Update TextRenderer to support color and alignment setters
- Standardize text alignment enum usage across components
- Add new UIComponent properties (clickable, scale)
- Move mouse input handling to InputManager
- Fix camera aspect ratio calculation

+++ src/com/odyssey/ui/layout/FrameLayout.java
@@ -207,8 +207,8 @@ public class FrameLayout extends LayoutManager {
                                 parent.getPaddingTop() + parent.getPaddingBottom());
         }
         
-        parent.setMinWidth(minWidth);
-        parent.setMinHeight(minHeight);
+        parent.setMinWidth((float)minWidth);
+        parent.setMinHeight((float)minHeight);
     }
     
     /**
